updated kafka message handling
