+++ eyeshade/models/models.go
@@ -1,15 +1,9 @@
 package models
 
 import (
-	"context"
-	"encoding/hex"
 	"errors"
-	"fmt"
-	"strconv"
 	"time"
 
-	"github.com/brave-intl/bat-go/utils/altcurrency"
-	"github.com/brave-intl/bat-go/utils/inputs"
 	"github.com/brave-intl/bat-go/utils/wallet/provider/uphold"
 	uuid "github.com/satori/go.uuid"
 	"github.com/shopspring/decimal"
@@ -68,297 +62,6 @@ type ConvertableTransaction interface {
 	Ignore() bool
 }
 
-// Settlement holds information from settlements queue
-type Settlement struct {
-	AltCurrency    altcurrency.AltCurrency
-	Probi          decimal.Decimal
-	Fees           decimal.Decimal
-	Amount         decimal.Decimal // amount in settlement currency
-	Currency       string
-	Owner          string
-	Channel        Channel
-	ID             string
-	Type           string
-	SettlementID   string
-	DocumentID     string
-	ExecutedAt     *time.Time
-	Address        string
-	WalletProvider *string
-}
-
-// CreatedAt computes the created at time of the settlement
-func (settlement Settlement) CreatedAt() time.Time {
-	createdAt := settlement.ExecutedAt
-	if createdAt != nil {
-		return *createdAt
-	}
-	parsed, err := time.Parse(time.RFC3339, settlement.ID)
-	if err == nil {
-		return parsed
-	}
-	id := settlement.ID
-	hexID := hex.EncodeToString([]byte(id))[:8]
-	parsedSec, err := strconv.ParseInt(hexID, 16, 64)
-	if err != nil {
-		return time.Unix(parsedSec, 0)
-	}
-	return time.Now()
-}
-
-// GenerateID generates an id from the settlement metadata
-func (settlement Settlement) GenerateID(category string) string {
-	return uuid.NewV5(
-		TransactionNS[category],
-		settlement.SettlementID+settlement.Channel.Normalize().String(),
-	).String()
-}
-
-// ToTxs converts a settlement to the appropriate transactions
-func (settlement Settlement) ToTxs() *[]Transaction {
-	txs := []Transaction{}
-	createdAt := settlement.CreatedAt()
-	month := createdAt.Month().String()[:3]
-	normalizedChannel := settlement.Channel.Normalize()
-	settlementType := settlement.Type
-	transactionType := settlementType + "_settlement"
-	toAccountType := "uphold"
-	if settlement.WalletProvider != nil {
-		toAccountType = *settlement.WalletProvider
-	}
-	if settlementType == "contribution" {
-		txs = append(txs, Transaction{
-			ID:              settlement.GenerateID("settlement_from_channel"),
-			CreatedAt:       createdAt.Add(time.Second * -2),
-			Description:     fmt.Sprintf("contributions through %s", month),
-			TransactionType: "contribution",
-			DocumentID:      settlement.ID,
-			FromAccount:     normalizedChannel.String(),
-			FromAccountType: "channel",
-			ToAccount:       settlement.Owner,
-			ToAccountType:   "owner",
-			Amount:          altcurrency.BAT.FromProbi(settlement.Probi.Add(settlement.Fees)),
-			Channel:         &normalizedChannel,
-		})
-		txs = append(txs, Transaction{
-			ID:              settlement.GenerateID("settlement_fees"),
-			CreatedAt:       createdAt.Add(time.Second * -1),
-			Description:     "settlement fees",
-			TransactionType: "fees",
-			DocumentID:      settlement.ID,
-			FromAccount:     settlement.Owner,
-			FromAccountType: "owner",
-			ToAccount:       "fees-account",
-			ToAccountType:   "internal",
-			Amount:          altcurrency.BAT.FromProbi(settlement.Fees),
-			Channel:         &normalizedChannel,
-		})
-	} else if settlementType == "manual" {
-		txs = append(txs, Transaction{
-			ID:              settlement.GenerateID(settlementType),
-			CreatedAt:       createdAt, // do we want to change this?
-			Description:     "handshake agreement with business development",
-			TransactionType: settlementType,
-			DocumentID:      settlement.DocumentID,
-			FromAccount:     SettlementAddress,
-			FromAccountType: toAccountType,
-			ToAccount:       settlement.Owner,
-			ToAccountType:   "owner",
-			Amount:          altcurrency.BAT.FromProbi(settlement.Probi),
-		})
-	}
-	txs = append(txs, Transaction{
-		ID:                 settlement.GenerateID(transactionType),
-		CreatedAt:          createdAt,
-		Description:        fmt.Sprintf("payout for %s", settlement.Type),
-		TransactionType:    transactionType,
-		FromAccount:        settlement.Owner,
-		FromAccountType:    "owner",
-		ToAccount:          settlement.Address,
-		ToAccountType:      toAccountType, // needs to be updated
-		Amount:             altcurrency.BAT.ToProbi(settlement.Probi),
-		SettlementCurrency: &settlement.Currency,
-		SettlementAmount:   &settlement.Amount,
-		Channel:            &normalizedChannel,
-	})
-	return &txs
-}
-
-// Ignore allows us to savely ignore a message if it is malformed
-func (settlement Settlement) Ignore() bool {
-	props := settlement.Channel.Normalize().Props()
-	return altcurrency.BAT.FromProbi(settlement.Probi).GreaterThan(largeBAT) ||
-		altcurrency.BAT.FromProbi(settlement.Fees).GreaterThan(largeBAT) ||
-		(props.ProviderName == "youtube" && props.ProviderSuffix == "user")
-}
-
-// Valid checks whether the message it has everything it needs to be inserted correctly
-func (settlement Settlement) Valid() bool {
-	// non zero and no decimals allowed
-	return !settlement.Probi.GreaterThan(decimal.Zero) &&
-		settlement.Probi.Equal(settlement.Probi.Truncate(0)) && // no decimals
-		settlement.Fees.GreaterThanOrEqual(decimal.Zero) &&
-		settlement.Owner != "" &&
-		settlement.Channel != "" &&
-		settlement.Amount.GreaterThan(decimal.Zero) &&
-		settlement.Currency != "" &&
-		settlement.Type != "" &&
-		settlement.Address != "" &&
-		settlement.DocumentID != "" &&
-		settlement.SettlementID != ""
-}
-
-// Referral holds information from referral queue
-type Referral struct {
-	AltCurrency   altcurrency.AltCurrency
-	Probi         decimal.Decimal
-	Channel       Channel
-	TransactionID string
-	Owner         string
-	FirstID       time.Time
-}
-
-// GenerateID creates a uuidv5 generated identifier from the referral data
-func (referral Referral) GenerateID() string {
-	return uuid.NewV5(
-		TransactionNS["referral"],
-		referral.TransactionID,
-	).String()
-}
-
-// ToTxs converts a referral to the appropriate transactions
-func (referral Referral) ToTxs() *[]Transaction {
-	month := referral.FirstID.Month().String()[:3]
-	normalizedChannel := referral.Channel.Normalize()
-	return &[]Transaction{{
-		ID:              referral.GenerateID(),
-		CreatedAt:       referral.FirstID,
-		Description:     fmt.Sprintf("referrals through %s", month),
-		TransactionType: "referral",
-		DocumentID:      referral.TransactionID,
-		FromAccount:     SettlementAddress,
-		FromAccountType: "uphold",
-		ToAccount:       referral.Owner,
-		ToAccountType:   "owner",
-		Amount:          altcurrency.BAT.FromProbi(referral.Probi),
-		Channel:         &normalizedChannel,
-	}}
-}
-
-// Ignore allows us to savely ignore the transaction if necessary
-func (referral Referral) Ignore() bool {
-	props := referral.Channel.Normalize().Props()
-	return altcurrency.BAT.FromProbi(referral.Probi).GreaterThan(largeBAT) ||
-		(props.ProviderName == "youtube" && props.ProviderSuffix == "user")
-}
-
-// Valid checks whether the transaction has the necessary data to be inserted correctly
-func (referral Referral) Valid() bool {
-	return referral.AltCurrency.IsValid() &&
-		referral.Probi.GreaterThan(decimal.Zero) &&
-		referral.Probi.Equal(referral.Probi.Truncate(0)) && // no decimals allowed
-		referral.Channel != "" &&
-		referral.Owner != "" &&
-		!referral.FirstID.IsZero()
-}
-
-// Votes holds information from votes freezing
-type Votes struct {
-	Amount            decimal.Decimal
-	Fees              decimal.Decimal
-	Channel           Channel
-	SurveyorID        string
-	SurveyorCreatedAt time.Time
-}
-
-// GenerateID generates an id from the contribution namespace
-func (votes Votes) GenerateID() string {
-	return uuid.NewV5(
-		TransactionNS["contribution"],
-		votes.SurveyorID+votes.Channel.String(),
-	).String()
-}
-
-// ToTxs converts votes to a list of transactions
-func (votes Votes) ToTxs() *[]Transaction {
-	return &[]Transaction{{
-		ID:              votes.GenerateID(),
-		CreatedAt:       votes.SurveyorCreatedAt,
-		Description:     fmt.Sprintf("votes from %s", votes.SurveyorID),
-		TransactionType: "contribution",
-		DocumentID:      votes.SurveyorID,
-		FromAccountType: "uphold",
-		ToAccount:       votes.Channel.String(),
-		ToAccountType:   "channel",
-		Amount:          votes.Amount.Add(votes.Fees),
-		Channel:         &votes.Channel,
-	}}
-}
-
-// Ignore allows us to ignore a vote if it is malformed
-func (votes Votes) Ignore() bool {
-	props := votes.Channel.Normalize().Props()
-	return votes.Amount.GreaterThan(largeBAT) ||
-		votes.Fees.GreaterThan(largeBAT) ||
-		(props.ProviderName == "youtube" && props.ProviderSuffix == "user")
-}
-
-// Valid checks that we have all of the information needed to insert the transaction
-func (votes Votes) Valid() bool {
-	return votes.Amount.GreaterThan(decimal.Zero) &&
-		votes.SurveyorID != "" &&
-		votes.Channel != "" &&
-		votes.Fees.GreaterThanOrEqual(decimal.Zero)
-}
-
-// UserDeposit holds information from user deposits
-type UserDeposit struct {
-	ID        string
-	Amount    decimal.Decimal
-	Chain     string
-	CardID    string
-	CreatedAt time.Time
-	Address   string
-}
-
-// GenerateID creates a uuidv5 generated identifier from the user deposit data
-func (userDeposit UserDeposit) GenerateID() string {
-	return uuid.NewV5(
-		TransactionNS["user_deposit"],
-		fmt.Sprintf("%s-%s", userDeposit.Chain, userDeposit.ID),
-	).String()
-}
-
-// ToTxs converts a user deposit to the appropriate transaction list
-func (userDeposit UserDeposit) ToTxs() *[]Transaction {
-	return &[]Transaction{{
-		ID:              userDeposit.GenerateID(),
-		CreatedAt:       userDeposit.CreatedAt,
-		Description:     fmt.Sprintf("deposits from %s chain", userDeposit.Chain),
-		TransactionType: "user_deposit",
-		DocumentID:      userDeposit.ID,
-		FromAccount:     userDeposit.Address,
-		FromAccountType: userDeposit.Chain,
-		ToAccount:       userDeposit.CardID,
-		ToAccountType:   "uphold",
-		Amount:          userDeposit.Amount,
-	}}
-}
-
-// Ignore allows us to savely ignore it
-func (userDeposit UserDeposit) Ignore() bool {
-	return userDeposit.Amount.GreaterThan(largeBAT)
-}
-
-// Valid checks that the transaction has everything it needs before being inserted into the db
-func (userDeposit UserDeposit) Valid() bool {
-	return userDeposit.CardID != "" &&
-		!userDeposit.CreatedAt.IsZero() &&
-		userDeposit.Amount.GreaterThan(decimal.Zero) &&
-		userDeposit.ID != "" &&
-		userDeposit.Chain != "" &&
-		userDeposit.Address != ""
-}
-
 // AccountEarnings holds results from querying account earnings
 type AccountEarnings struct {
 	Channel   Channel         `json:"channel" db:"channel"`
@@ -394,153 +97,4 @@ type Balance struct {
 	AccountID string          `json:"account_id" db:"account_id"`
 	Type      string          `json:"account_type" db:"account_type"`
 	Balance   decimal.Decimal `json:"balance" db:"balance"`
-}
-
-// PendingTransaction holds information about an account id's pending transactions
-type PendingTransaction struct {
-	Channel Channel         `db:"channel"`
-	Balance decimal.Decimal `db:"balance"`
-}
-
-// Transaction holds info about a single transaction from the database
-type Transaction struct {
-	ID                 string           `db:"id"`
-	CreatedAt          time.Time        `db:"created_at"`
-	Description        string           `db:"description"`
-	TransactionType    string           `db:"transaction_type"`
-	DocumentID         string           `db:"document_id"`
-	FromAccount        string           `db:"from_account"`
-	FromAccountType    string           `db:"from_account_type"`
-	ToAccount          string           `db:"to_account"`
-	ToAccountType      string           `db:"to_account_type"`
-	Amount             decimal.Decimal  `db:"amount"`
-	SettlementCurrency *string          `db:"settlement_currency"`
-	SettlementAmount   *decimal.Decimal `db:"settlement_amount"`
-	Channel            *Channel         `db:"channel"`
-}
-
-// BackfillForCreators converts a transaction from the database to a backfill transaction
-func (tx Transaction) BackfillForCreators(account string) CreatorsTransaction {
-	amount := tx.Amount
-	if tx.FromAccount == account {
-		amount = amount.Neg()
-	}
-	var settlementDestinationType *string
-	var settlementDestination *string
-	if SettlementTypes[tx.TransactionType] {
-		if tx.ToAccountType != "" {
-			settlementDestinationType = &tx.ToAccountType
-		}
-		if tx.ToAccount != "" {
-			settlementDestination = &tx.ToAccount
-		}
-	}
-	inputAmount := inputs.NewDecimal(&amount)
-	inputSettlementAmount := inputs.NewDecimal(tx.SettlementAmount)
-	return CreatorsTransaction{
-		Amount:                    inputAmount,
-		Channel:                   tx.Channel,
-		CreatedAt:                 tx.CreatedAt,
-		Description:               tx.Description,
-		SettlementCurrency:        tx.SettlementCurrency,
-		SettlementAmount:          inputSettlementAmount,
-		TransactionType:           tx.TransactionType,
-		SettlementDestinationType: settlementDestinationType,
-		SettlementDestination:     settlementDestination,
-	}
-}
-
-// CreatorsTransaction holds a backfilled version of the transaction
-type CreatorsTransaction struct {
-	CreatedAt                 time.Time      `json:"created_at"`
-	Description               string         `json:"description"`
-	Channel                   *Channel       `json:"channel"`
-	Amount                    inputs.Decimal `json:"amount"`
-	TransactionType           string         `json:"transaction_type"`
-	SettlementCurrency        *string        `json:"settlement_currency,omitempty"`
-	SettlementAmount          inputs.Decimal `json:"settlement_amount,omitempty"`
-	SettlementDestinationType *string        `json:"settlement_destination_type,omitempty"`
-	SettlementDestination     *string        `json:"settlement_destination,omitempty"`
-}
-
-// SettlementStat holds settlement stats
-type SettlementStat struct {
-	Amount inputs.Decimal `json:"amount" db:"amount"`
-}
-
-// SettlementStatOptions holds options for scaling up and down settlement stats
-type SettlementStatOptions struct {
-	Type     string
-	Start    inputs.Time
-	Until    inputs.Time
-	Currency *string
-}
-
-// NewSettlementStatOptions creates a bundle of settlement stat options
-func NewSettlementStatOptions(
-	t, currencyInput, layout, start, until string,
-) (*SettlementStatOptions, error) {
-	startTime, untilTime, err := parseTimes(layout, start, until)
-	if err != nil {
-		return nil, err
-	}
-	var currency *string
-	if currencyInput != "" {
-		currency = &currencyInput
-	}
-	return &SettlementStatOptions{
-		Type:     t,
-		Start:    *startTime,
-		Until:    *untilTime,
-		Currency: currency,
-	}, nil
-}
-
-// GrantStat holds Grant stats
-type GrantStat struct {
-	Amount inputs.Decimal `json:"amount" db:"amount"`
-	Count  inputs.Decimal `json:"count" db:"count"`
-}
-
-// GrantStatOptions holds options for scaling up and down grant stats
-type GrantStatOptions struct {
-	Type  string
-	Start inputs.Time
-	Until inputs.Time
-}
-
-// NewGrantStatOptions creates a new grant stat options object
-func NewGrantStatOptions(
-	t, layout, start, until string,
-) (*GrantStatOptions, error) {
-	startTime, untilTime, err := parseTimes(layout, start, until)
-	if err != nil {
-		return nil, err
-	}
-	if !ValidGrantStatTypes[t] {
-		return nil, fmt.Errorf("unable to check grant stats for type %s", t)
-	}
-	return &GrantStatOptions{
-		Type:  t,
-		Start: *startTime,
-		Until: *untilTime,
-	}, nil
-}
-
-func parseTimes(
-	layout, start, until string,
-) (*inputs.Time, *inputs.Time, error) {
-	var (
-		startTime = inputs.NewTime(layout)
-		untilTime = inputs.NewTime(layout)
-		ctx       = context.Background()
-	)
-	if err := startTime.Decode(ctx, []byte(start)); err != nil {
-		return nil, nil, err
-	}
-	if err := untilTime.Decode(ctx, []byte(until)); err != nil {
-		untilSubMonth := startTime.Time().AddDate(0, -1, 0)
-		untilTime.SetTime(untilSubMonth)
-	}
-	return startTime, untilTime, nil
 }
